aclconverter_0_15.pl: add URG flag to "established"
I have found a TCP trace containing a packet with the URG flag set.
This packet is part of a TELNET session, and it has flags PSH and ACK
set as well. This shows that at least the combination of the three
flags PSH, URG, and ACK need to be allowed as part of an "established"
connection.

The combination URG and ACK (without PSH) should be considered valid
as well, since URG and PSH are separate mechanisms. As such it seems
best to allow that combination as well.

RFC 6093, On the TCP Urgent Mechanism, gives two example applications
that use the URG flag: TELNET and FTP. In section 3.4. it states that
"urgent indications may not be reliable in the current Internet".

The addition of URG+ACK and PSH+URG+ACK to translation of "established"
aims to create a useful translation. A 1-to-1 translation would need
to specify all possible flag combinations containing ACK, RST, or both
flags, which I do not consider useful.