Fix for Issue "Mlag check script #30"
Script was incorrectly handling MLAG ports with no VLANs added.  The user was presented with a warning and the script incorrectly suggested to add VLAN "None" to the ISC.  The script now checks for empty VLAN membership on a MLAG port and presents a warning informing the user there is no VLAN configuration on the port.  The suggestion to add VLAN "None" to the ISC has been removed